+++ scripts/rails.js
